pyproject.toml
[project]
name = "bd-project-25"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dask-geopandas<0.5.0",
    "dask-jobqueue>=0.9.0",
    "dask-sql>=2024.5.0",
    "dask[complete]<=2025.0.1",
    "duckdb>=1.2.2",
    "folium>=0.19.5",
    "geopandas>=1.0.1",
    "h5py>=3.13.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "matplotlib>=3.10.3",
    "pyarrow>=19.0.1",
    "scipy>=1.16.0",
    "seaborn>=0.13.2",
    "tables>=3.10.2",
    "tqdm>=4.67.1",
]
